app/src/main/java/com/arya/storyapp/ui/viewmodel/MainViewModel.kt
package com.arya.storyapp.ui.viewmodel

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.arya.storyapp.model.Story
import com.arya.storyapp.repository.StoryRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import javax.inject.Inject

@HiltViewModel
class MainViewModel @Inject constructor(
    private val storyRepository: StoryRepository,
) : ViewModel() {
    fun getAllStories(): LiveData<PagingData<Story>> {
        return storyRepository.getAllStories().cachedIn(viewModelScope)
    }
}

